Move /v1 routes in deviceState.router to v1.ts
