+++ test/lib/mocked-device-api.ts
@@ -132,7 +132,7 @@ async function create(
 
 	// Create SupervisorAPI
 	const api = new SupervisorAPI({
-		routers: [v1.router, deviceState.router],
+		routers: [v1.router],
 		healthchecks,
 	});
